+++ src/main/java/cart/controller/AdminController.java
@@ -3,8 +3,8 @@ package cart.controller;
 import cart.domain.product.dto.ProductCreateDto;
 import cart.domain.product.dto.ProductUpdateDto;
 import cart.domain.product.service.ProductService;
-import cart.dto.ProductCreateRequest;
-import cart.dto.ProductUpdateRequest;
+import cart.dto.request.ProductCreateRequest;
+import cart.dto.request.ProductUpdateRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
