+++ src/main/java/cart/controller/MemberController.java
@@ -4,8 +4,8 @@ import cart.auth.EncodePassword;
 import cart.domain.member.dto.CreatedMemberDto;
 import cart.domain.member.dto.MemberDto;
 import cart.domain.member.service.MemberService;
-import cart.dto.MemberCreateRequest;
-import cart.dto.MemberCreateResponse;
+import cart.dto.request.MemberCreateRequest;
+import cart.dto.response.MemberCreateResponse;
 import javax.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
