+++ src/main/java/cart/controller/ViewController.java
@@ -4,8 +4,8 @@ import cart.domain.member.dto.MemberDto;
 import cart.domain.member.service.MemberService;
 import cart.domain.product.dto.ProductDto;
 import cart.domain.product.service.ProductService;
-import cart.dto.MemberResponse;
-import cart.dto.ProductResponse;
+import cart.dto.response.MemberResponse;
+import cart.dto.response.ProductResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.http.ResponseEntity;
