+++ src/main/java/cart/dto/ExceptionDto.java
@@ -1,4 +1,4 @@
-package cart.exception;
+package cart.dto;
 
 public class ExceptionDto {
