+++ src/main/java/cart/dto/request/MemberCreateRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 import javax.validation.constraints.Email;
