+++ src/main/java/cart/dto/request/ProductCreateRequest.java
@@ -1,12 +1,10 @@
-package cart.dto;
+package cart.dto.request;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.PositiveOrZero;
 
-public class ProductUpdateRequest {
+public class ProductCreateRequest {
 
-    @NotBlank
-    private Long id;
     @NotBlank
     private String name;
     @PositiveOrZero(message = "가격은 음수가 될 수 없습니다.")
@@ -14,15 +12,13 @@ public class ProductUpdateRequest {
     @NotBlank
     private String imageUrl;
 
-    public ProductUpdateRequest() {
+    public ProductCreateRequest() {
     }
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(final Long id) {
-        this.id = id;
+    public ProductCreateRequest(final String name, final int price, final String imageUrl) {
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
     }
 
     public String getName() {
