+++ src/main/java/cart/dto/request/ProductUpdateRequest.java
@@ -1,10 +1,12 @@
-package cart.dto;
+package cart.dto.request;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.PositiveOrZero;
 
-public class ProductCreateRequest {
+public class ProductUpdateRequest {
 
+    @NotBlank
+    private Long id;
     @NotBlank
     private String name;
     @PositiveOrZero(message = "가격은 음수가 될 수 없습니다.")
@@ -12,13 +14,15 @@ public class ProductCreateRequest {
     @NotBlank
     private String imageUrl;
 
-    public ProductCreateRequest() {
+    public ProductUpdateRequest() {
     }
 
-    public ProductCreateRequest(final String name, final int price, final String imageUrl) {
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(final Long id) {
+        this.id = id;
     }
 
     public String getName() {
