+++ src/main/java/cart/dto/response/CartResponse.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.response;
 
 import cart.domain.cart.dto.CartDto;
