+++ src/test/java/cart/controller/AdminControllerTest.java
@@ -7,7 +7,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import cart.domain.product.service.ProductService;
-import cart.dto.ProductCreateRequest;
+import cart.dto.request.ProductCreateRequest;
 import cart.domain.product.dto.ProductUpdateDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
