+++ src/test/java/cart/controller/CartControllerTest.java
@@ -13,7 +13,7 @@ import cart.auth.AuthInfoArgumentResolver;
 import cart.auth.AuthService;
 import cart.auth.AuthenticationCheckInterceptor;
 import cart.domain.cart.dto.CartDto;
-import cart.dto.CartResponse;
+import cart.dto.response.CartResponse;
 import cart.domain.cart.service.CartService;
 import cart.domain.member.dto.MemberDto;
 import cart.domain.product.dto.ProductDto;
